Create MentorCheckBox-handleChange
The handleChange function handles the change, and if the checkbox is checked, the value would equal true.  If the value is true then make a POST AJAX call to API and return the id.  or else if it's false, then do DELETE AJAX call to API using the return id.